+++ src/index.ts
@@ -16,5 +16,6 @@ export * from './definitions';
 export * from './core/enums';
 export * from './core/models';
 export * from './core/options';
+export * from './core/text';
 export * from './core/errors';
 export { ThermalPrinter };